move record ops to its own package
